requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
pandas>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
numpy>=1.26.0

